// src/models/category.enum.ts
export enum CategoryTitle {
  generalColors = "General colors",
  globalSizes = "Global sizes",
  textField = "Text field",
  buttons = "Buttons",
}

export enum Category {
  generalColors = "generalColors",
  globalSizes = "globalSizes",
  textField = "textField",
  buttons = "buttons",
}

export enum CategoryProperty {
  border = "Border",
  fontSizeRem = "Font size",
}